Fix warnings

+++ libs/oef-search/include/oef-search/comms/Oefv1Listener.hpp
@@ -20,7 +20,7 @@ public:
 
   static constexpr char const *LOGGING_NAME = "Oefv1Listener";
 
-  Oefv1Listener(std::shared_ptr<Core> core, int port, ConfigMap endpointConfig);
+  Oefv1Listener(std::shared_ptr<Core> core, unsigned short int port, ConfigMap endpointConfig);
   virtual ~Oefv1Listener()
   {
     std::cout << "Listener on " << port << " GONE" << std::endl;
